Created calibration function, writing data output to .csv
created a function that calibrates the centrepoint by averaging the min
and max of the magnetic field at all three axis. This calibration is
then subtracted from all the future test carried out, until a new
calibration is made. Some adjustments need to be made as the y axis does
not seem to be changing by a large factor. Possible need to rewrite the
code alltogether if it's a hardware sampling issue